Added VARSET MOBILE WORLD alias to VARSET MOBILE NAME for symmetry.
Added VARSET MOBILE HERE for symmetry.
Added VARSET OBJECT NAME alias to VARSET OBJECT WORLD for symmetry.
Removed the viewer requirement for searching mobiles and objects in as much as possible in VARSET.
Will have to add an optional argument to specify the viewer of the search, for cases where that is required.