Added ENTRY trigger usage for dungeons and instances.
Added REPOP trigger usage for dungeons and instances.
 - instance REPOPs are only fired by the controlling entity, such as dungeons.
Added RESET trigger usage for rooms, areas, dungeons and instances.
 - When an instance is part of the dungeon, it will not fire any RESET triggers of its own.
Added reset interval for dungeons and instances.
IF LOADED now accepts INSTANCE and DUNGEON entities after MOBILE info to check the mob's count in the INSTANCE/DUNGEON.
Added DUNGEONCOMPLETE and INSTANCECOMPLETE for firing their COMPLETED trigger.
Added MLOAD and OLOAD to AREA, DUNGEON, INSTANCE progs
Added MAKEINSTANCED to DUNGEON and INSTANCE progs, for making mobiles and objects that are loaded to be instance mobs/objs.  Use when loading mobiles and objects by triggers so they are treated as part of the instance.  Only mobs/objs not already flagged and currently IN the instance or dungeon calling the command.