Restored key usage (for lock, unlock and pick) for portals.
Moved the object value used to specify the AREA number for portals on random to v5, since the other use for v5 (wilderness portals) is mutually exclusive from the intended use for the AREA number on portals.  Now the Area Number will be used with AREARANDOM portal flag to specify the area to go to.
Dungeon portals now use v5 (instead of v4) for the floor designation, to allow for keys.
Books in OEDIT can be given container flags and a key definition.
Books can now be opened, closed, locked, unlocked and picked!