Areas lacked a uid identifying what WILDERNESS they were in, if any.
Airships should be able to land into and launch from areas, provided they meet the requirements for the airships.
Landing requires the airship be within 20 blocks of an area's X/Y and in the same wilderness.  And the specified area requires an airship landing spot defined.
Launching from an area requires the area have a defined Wilderness Vnum and an X/Y
Moving a ship failed to check if the new coordinate was still on the map.  Need to add the ability to wrap around the map if defined properly.