Widevnum changes (compile test only)
Updated things to use widevnums.  Everything is under areas that needs to be.
Moved blueprints, dungeons and ships under areas.  Will add
Reworked portals some to have different portal types for handling different types of destinations.
Removed some old functions and commands that either need to be gone or reworked entirely.
Added a bootstrap process when the system detects no loaded areas.  It will create a dummy area and room to start the building process, updating the reserved rooms to said starting room.  The player that initiates the process will automatically become an IMP with full security.
Expanded the number of object and token values to 10.
Added reserved entity handling, like rooms, so they are loadable and configurable in-game.
Made churches where they can be in any area designated with AREA_CHURCH.  Prepping functionality to allow them to relocate to different areas wholesale.