Fixed a crash bug
When going to Limbo from a VRoom and subsequently idle quitting, it was not saving the correct information to the pfile.
Added a check on login when returning to a VRoom to see if it is still a valid room, as that can be changed between logins.  If it is bad, it will send you back to the TEMPLE reserved room.
Added configuration settings in gconfig for limbo timeout and disconnect timeouts, defaulting to 12 and 30 minutes respectively.  The disconnect must be larger than the limbo timeout.