Death release modes (compile only)
Added support for the death release modes for dungeons.
 - Dying in a dungeon will not teleport you to the death plane unless the dungeon uses normal death releasing.
Added DUNGEON RELEASE command for when you are dead inside a dungeon.
Added failure state for dungeons and instances.
Added FAILED trigger for dungeon and instance progs.
Added DUNGEONFAILURE and INSTANCEFAILURE script commands.
Made ships and treasure maps check all of an area's regions when determining which one is closest to a coordinate on a map.
Mob, Obj, and Room stat commands will show whether they are persistent.