Custom trigger types (part 2)
Added support for making custom script triggers, basically the codification of the xtrigger system in the legacy server.
Added script commands for all of the trigger functions for use in scripts for nefarious (or very nefarious) purposes.
Fixed the loading of scripts from AREA files to update the top script index.
STAT TOKEN will now show the widevnum of the tokens, instead of just the vnum.
Added a usage counter for trigger types (used in determining whether a custom trigger type may be deleted.)