Added ability to install new builtin triggers
Restricted the triggers command to those with security 9 for the time being, until the role system is put in place where we can have those with the appropriate role.
Modifying builtin triggers is now possible for updating their script spaces.  Security 10 will be required for those that are in use.