Added AUTOOLC toggle for building
When on, REDIT and REDIT <widevnum> will save the current area, region, sector, and room flags.  Changes made to any of those on the room will update the saved settings.  If you create a new room that is outside the saved area, it will reset, requiring the builder to make changes manually.  If it is the same area, it will apply those settings to the new room.