Changes to fountains, drink containers, food and poisoning.
Also added four triggers:
 EMPTIED (works on all containers and fountains)
 FILL (works on fountains and drinks)
 FILLED (works on all containers and fountains)
 POUR (works on fountains and drinks)

Poisoning fountains and drinks is no longer permanent as it will erode with use, especially when emptying out and refilling.

Fountains are not distinct from drinks in that they have a FILL RATE value (v4) that specifies how much they fill up each tick, allowing them to automatically refill.

Fixed the TRIGGER INSTALL command to get the correct script spaces.

Food, Drinks and Fountains will erode the poison that is in them, eventually.  Occasionally, the poison will completely dissipate.

Consuming anything poisoned will attempt to inflict POISON with a duration based upon the poison's percent value along with the amount consumed.

Fixed an issue with PUT ALL/ALL.object where it would double message.  One of the messages was not obeying the silent flag, so it was posting regardless.

The POISON spell will now work on FOUNTAINs.

OSTAT will not show all 10 object values.
The IDENTIFY spell/LORE ability for DRINKs will now indicate whether it is empty instead of stating it has something in it.