Converted multiple flag sets into "bitmatrices"
Reworked the BIT ifcheck to take a BITVECTOR or a BITMATRIX entity and check if all of the specified bits are set.
Deprecated the ACT/2, AFFECTED/2 and OBJEXTRA/2/3/4 ifchecks in favor of using the BIT ifcheck with the appropriate bitvector/matrix field.
BITVECTOR and BITMATRIX fields will now generate their string outputs when used in string expansions.