Various bug fixes
Varseton and varclearon were not targeting $AREA, $INSTANCE and $DUNGEON types in the mprog, oprog, rprog and tprog versions.

MOBLIST and OBJLIST varset/on was missing ability to remove a specific mobile/object from their list.

Varset/on was not allowing the variable type to be an expansion (such as a variable) as it was requiring it to be a literal string.

Aedit and Redit "should" now update both their index variables and their live versions when using the varset/varclear commands in the editor.  For rooms, though, this will *only* affect the static room.  Clone rooms will not be updated.