WAND multityping, MXP support started and more
WAND multityping is incomplete, and some aspects, such as ZAP need some work, pending completion of the spell deflection updates.

Started MXP support as well as Tabbed editors, starting with SKEDIT.

Spell deflection is getting a small rework, to make it both a bit simpler for different actions handling it as well as able to be customizied by the actual spell.