Start of WEAPON and AMMO multityping.
Combined ranged weapons with WEAPON.

Ammo can be generated using AMMO_SCRIPTED trigger on the ranged weapon.  To generate ammo, I added the SETOUTBOUND script command.  Currently, I only allow objects and tokens on objects that are calling the AMMO_SCRIPTED trigger to be able to use this command, as there aren't any other situations that need this functionality yet.

The script command OLOAD now sets a CREATED flag on the object that.  This is primarily used by the ranged code when interrupting ranged to prevent generated ammo from piling up when the action is interrupted.  Should the action make it to the ranged_end function or the projectile is originally pulled from an inventory, that flag is removed automatically, so as not to be pruned in error.

Ranged weapons can use mana to generate ammo.  Use the AMMO_MANA trigger (tempstore1) to set the mana cost.

Added EPHEMERAL object flag that causes the item to be consumed after use (currently only the shooting code checks this) or will cause the item to vanish outside of social on the next game tick.

Still have the OLC component to finish.

Added HOSTILE flag on reputation ranks.  Mobs associated with reputations where a target is at a hostile rank (such as a player) will cause the mob to attack them, as if they were aggressive.  This cannot be stopped with the PEACE command, however.  That can be changed, though.

Added the ability to set the BRANDISH function on skills now, in prep for being able to make OLC weapons that have spells on them, and, later, the entire imbuing system.

MAJOR REWORK NEEDED: weapon use in combat and combat related things.  This can be addressed in the combat update, as we may need to alter weapons more.