Added ARMOR multityping with ADORNMENTS.  Compile test only.
Only the definitions and OLC are done.  Actually using them still needs to be done.

Adornments, regardless of type, can only have one equipable spell on them, and must be unique across the entire armor.  This is true in OLC and will be so when trying to add said adornments as a player.