Coupled tokens and affects.
Affects can be associated with AFFECT tokens such that they are aware of each other when one is removed.  When an affect token is removed, all affects associated with it as removed from its environment as if they wore off.

Expanded the GET trigger usage so that mob scripts can handle getting an object.  It is not perfect, due to how the GET command operates.

Fixed an issue with many numerical triggers where $(OBJ2) was assigned $(OBJ1).