Stuff
Added a description field to all entity fields for a future helper command that allows a scripter to determine what $ENTITY expansion to use.
Changed the church key field in CHURCH_DATA to an OBJ_INDEX_DATA * instead of being a vnum, to keep from having to look it up every time we want the key object.
Modified a few field types, primarily to be STAT instead of a NUMBER or STRING.