Added hint system for dialogue choices
Choice visibility scripts have been changed to check three return values: 0/allowed, 1/denied and 2/silent.
allowed = visible and enabled
denied = visible and disabled
silent = hidden

For denied to work, the choice option MUST have a HINT field assigned to it.  It can be multi-lined and will be indented when shown.

The HINT field is ONLY copied over if the option's visibility script returned denied.  Otherwise, it will not be copied over.  Attempting to select a choice with a valid HINT field will be considered a bad selection.

Fixed the IS_NUMBER function to return false if nothing follows the sign character.